third_party/blink/web_tests/external/wpt/css/css-shapes/spec-examples/shape-outside-018-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: shape-margin offset from a polygonal shape-outside assert_approx_equals: Line 0 is positioned properly expected 48 +/- 2 but got 180
Harness: the test ran to completion.
